requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
anytree==2.12.1
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
asyncmy==0.2.9
attrs==24.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.1
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
imbalanced-learn==0.12.3
import-ipynb==0.1.4
importlib_metadata==7.2.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdb_parser==0.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic-settings==2.4.0
pydantic_core==2.18.4
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.4
scikeras==0.13.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
setuptools==70.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.2
tenacity==8.4.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wordcloud==1.9.3
wrapt==1.16.0
zipp==3.19.2
